pyproject.toml
[project]
name = "ccc"
version = "0.1.0"
description = "Cursed Content Creator"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "nicegui>=2.24.1",
    "numpy>=2.3.3",
    "pillow>=11.3.0",
    "reflex>=0.8.7",
    "requests>=2.32.5",
]

[dependency-groups]
dev = [
    "pyright>=1.1.405",
    "pytest>=8.4.2",
]
